+++ Runtime/Framework/CommandBus/CommandBus.cs
@@ -3,37 +3,35 @@ using System.Collections.Generic;
 using UnityEngine.Events;
 
 namespace DT.UniStart {
-  public class CommandBus : IWritableCommandBus {
+  public class CommandBus : ICommandBus {
     readonly IEventBus bus;
-    readonly Dictionary<Type, object> commands;
+    readonly HashSet<Type> commands = new();
 
     public CommandBus(IEventBus bus = null) {
       this.bus = bus ?? new EventBus();
-      this.commands = new();
     }
 
-    public UnityAction Add<T>(UnityAction command) where T : ICommand {
-      // IMPORTANT! transform UnityAction into UnityAction<T> for latter Get
-      this.commands.Add(typeof(T), UniStart.Fn((T _) => command.Invoke()));
+    public virtual UnityAction Add<T>(UnityAction command) where T : ICommand {
+      this.CheckCommand<T>();
       this.bus.AddListener<T>(command);
       return command;
     }
-    public UnityAction<T> Add<T>(UnityAction<T> command) where T : ICommand {
-      this.commands.Add(typeof(T), command);
+    public virtual UnityAction<T> Add<T>(UnityAction<T> command) where T : ICommand {
+      this.CheckCommand<T>();
       this.bus.AddListener(command);
       return command;
     }
 
-    public UnityAction<T> Get<T>() where T : ICommand => this.commands.GetOrDefault(typeof(T)) as UnityAction<T>;
+    public virtual void Push<T>(T arg) where T : ICommand => this.bus.Invoke(arg);
 
-    public void Push<T>(T arg) where T : ICommand => this.bus.Invoke(arg);
+    void CheckCommand<T>() {
+      if (!this.commands.Add(typeof(T)))
+        throw new InvalidOperationException($"Command of type {typeof(T)} already exists!");
+    }
   }
 
-  // helper classes
+  // TODO: move to a single file
   public class DebugCommandBus : CommandBus {
-    public DebugCommandBus(InterceptEventBusMode mode = InterceptEventBusMode.Invoke) : base(new DebugEventBus(name: "DebugCommandBus", mode: mode)) { }
-  }
-  public class DelayedCommandBus : CommandBus {
-    public DelayedCommandBus() : base(new DelayedEventBus()) { }
+    public DebugCommandBus(DebugEventBusMode mode = DebugEventBusMode.Invoke) : base(new DebugEventBus(mode: mode).WithName("DebugCommandBus")) { }
   }
 }